rewrite it to it's equivalent in ECMAScript notation
